Clean up u_long/u_short/u_char types

+++ Source/3D/3DMath.c
@@ -1403,7 +1403,7 @@ void OGLMatrix4x4_Transpose(const OGLMatrix4x4 *matrix4x4, OGLMatrix4x4 *result)
 {
 OGLMatrix4x4 		source;
 const OGLMatrix4x4	*sourcePtr;
-u_long 				row, column;
+uint32_t 				row, column;
 	
 	if (result == matrix4x4)
 	{
@@ -2513,8 +2513,8 @@ float			lX = 0, lY = 0, lZ = 0;				// Local space co-ordinates
 float			hX = 0, hY = 0, hZ = 0, hW = 0;		// Homogeneous co-ordinates
 float			minusHW;				// -hW
 
-u_long		clipFlags;				// Clip in/out tests for point
-u_long		clipCodeAND;			// Clip test for entire object
+uint32_t		clipFlags;				// Clip in/out tests for point
+uint32_t		clipCodeAND;			// Clip test for entire object
 
 long			i;
 float			minX,minY,minZ,maxX,maxY,maxZ;
